backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.1.0,<4.2.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.30",
    "uvicorn>=0.40.0",
    "email-validator>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
]
